Refactor search and replace functionality
